save state - trying to get mongo working
