+++ node/40_41/slasherapi/.env
@@ -1,0 +1,1 @@
+MONGOPASS="AveryBerman!3!"
