+++ node/40_41/slasherapi/.gitignore
@@ -1,0 +1,2 @@
+node_modules
+dotenv
