+++ node/40_41/slasherapi/index.html
@@ -8,5 +8,11 @@
 <body>
   <h1>My API</h1>
   <p>Please use localhost:8000/api/myapi</p>
+
+  <form action="/slashers" method="POST">
+    <input type="text" placeholder="name" name="birthName" />
+    <input type="text" placeholder="birthplace" name="birthLocation" />
+    <button type="submit">Submit</button>
+  </form>
 </body>
 </html>
